refactor(core,addon,blocks): consolidate dataAttr onto core/data-attr subpath
Third subpath in the same series as /css-var (#880) and the existing
/resolve-at, /fuzzy. dataAttr lived in three places with identical
bodies (core/src/css-axis-projected.ts inline, addon/src/data-attr.ts
standalone, blocks/src/internal/data-attr.ts alongside themeAttrs).

Consolidates onto a new browser-safe core/data-attr subpath. addon's
standalone file is deleted; blocks/internal/data-attr.ts keeps
themeAttrs/BLOCK_ATTR/WRAPPER_CLASSES but imports dataAttr from core.
AxisVariance.tsx (the only blocks consumer that imported dataAttr
directly) switches to importing from the core subpath; previously it
reached through blocks/internal/data-attr.ts.

data-attr.test.ts moves from addon/test/ to core/test/ — follows the
impl.

Closes the audit's #824 consolidation cluster (canonicalKey/valueKey/
cssEscape in #869, findPermByTuple eliminated by cartesian drop,
makeCssVar in #880, dataAttr here). defaultTuple builder explicitly
skipped — 2-line inlines, extraction would add an import for no win.
