+++ packages/addon/src/preview.tsx
@@ -8,7 +8,7 @@ import type {
 import type { Decorator, Preview } from '@storybook/react-vite';
 import { useEffect, useMemo } from 'react';
 import { addons } from 'storybook/preview-api';
-import { dataAttr } from '#/data-attr.ts';
+import { dataAttr } from '@unpunnyfuns/swatchbook-core/data-attr';
 // Side-effect import for integrations that opted into `autoInject`
 // (e.g. Tailwind's `@theme` block). When no integration opts in, the
 // virtual module body is empty — still a valid no-op.
