+++ packages/core/test/data-attr.test.ts
